Temp branch/commit to fix the docs changelog for 1.6.1.
This commit represents the hash which the docs were previously on
(release-1.6.0.1) but with the exact History.md for Meteor 1.6.1 in
order to properly generate the "Changelog" which is shown on the
sidebar.

This is necessary because some of the jsdoc declarations for changes in
the 1.6.1 release affected the docs generation, which are expecting
particular declarations which are apparently missing.

Need to look into it more but hoping to avoid our usual changelog
confusion from the community. :)

cc @benjamn @hwillson